pyproject.toml
[project]
name = "unstopscrapper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.52.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
